Remove TiccleStackNavigator
