+++ src/navigation/mainTabComponent/TiccleCreateHeaderLeft.js
@@ -1,0 +1,34 @@
+import React from 'react';
+import { Image,StyleSheet,TouchableOpacity } from 'react-native';
+import { useNavigation } from '@react-navigation/native';
+import colors from '../../theme/colors';
+
+const TiccleCreateHeaderLeft = () => {
+    const navigateTo = useNavigation()
+    return (
+        <TouchableOpacity
+            style={styles.headerLeftTouchable}
+            onPress={() => navigateTo.navigate('HomeStack')}>
+                <Image source={require('../../assets/images/chevron_left.png')}
+                        style={styles.headerLeftImage}
+                />
+        </TouchableOpacity>
+    )
+}
+
+const styles = StyleSheet.create({
+    headerLeftTouchable :{
+        alignItems: 'center',
+        justifyContent : 'center',
+        width: 40,
+        height : 40
+    },
+    headerLeftImage : {
+        resizeMode : 'cover',
+        width : 12, 
+        height: 20, 
+        tintColor : colors.white
+    },
+})
+
+export default TiccleCreateHeaderLeft
